pyproject.toml
[project]
name = "contact_service"
version = "0.1.0"
description = ""
authors = [
    {name = "TarnavskyAndrew",email = "87767534+TarnavskyAndrew@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "bcrypt (>=4.0.1,<5.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "redis (>=6.4.0,<7.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "sphinx (>=8.2.3,<9.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-mock (>=3.15.0,<4.0.0)",
    "httpx (>=0.27)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = ">=0.24"
sphinx = "^8.2.3"
furo = "^2025.7.19"
aiosqlite = "^0.21.0"
freezegun = "^1.5.5"


[tool.poetry.group.test.dependencies]
httpx = "^0.28.1"
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
pytest-cov = "^7.0.0"
requests = "^2.32.5"

[tool.pytest.ini_options]
pythonpath = ["."]